requirements.txt
fastapi
uvicorn
pydantic
rapidfuzz
twilio
python-dotenv
requests
psycopg2-binary
redis
